pyproject.toml
[build-system]
requires = ['setuptools>=40.8.0']
build-backend = 'setuptools.build_meta'


[project]
name = "abdm-python-integrator"
description = "A Django app that implements ABDM milestones"
readme = "README.md"
keywords = [
    "abdm"
]
license = {text = "BSD-3-Clause"}
classifiers = [
    "Framework :: Django",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers"
]
requires-python = ">=3.8"
dependencies = [
    "django >= 3.2",
    "djangorestframework >= 3.12",
    "requests",
    "drf-standardized-errors",
    "celery",
    "fidelius @ git+https://github.com/dimagi/pyfidelius.git@master",
    "jsonpath-ng @ https://github.com/kaapstorm/python-jsonpath-rw/raw/wherenot+find_or_create/wheel/jsonpath_ng-1.5.2.2-py3-none-any.whl",
    "pyjwt[crypto]"
]
dynamic = ["version"]


[project.optional-dependencies]
code-quality = ["flake8", "flake8-isort"]

[tool.setuptools.packages.find]
include = ["abdm_integrator*"]
exclude = ["*tests*"]

[tool.setuptools.package-data]
"abdm_integrator.hiu.fhir" = ["*.json"]

[tool.isort]
line_length = 115
known_first_party = [
    "abdm-python-integrator",
]
skip = ["venv/"]
skip_glob = ["**/migrations/*.py"]
multi_line_output=3
include_trailing_comma=true
